requirements.txt
aiocron==1.8
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
croniter==2.0.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
kombu==5.3.5
magic-filter==1.0.12
multidict==6.0.5
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.8
pytz==2024.1
redis==5.0.2
requests==2.31.0
schedule==1.2.1
six==1.16.0
sniffio==1.3.1
telegram==0.0.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
